zram: add config init/release backend callbacks
Some backends can create a backend-specific private data for comp config,
e.g.  parse a dictionary and use it to init all of the ctx-s later on. 
Introduce two zcomp_config callbacks to create and destroy per-config
backend private data.  This is also the place where config can be
validated.

Link: https://lkml.kernel.org/r/20240515071645.1788128-17-senozhatsky@chromium.org
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Nick Terrell <terrelln@fb.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>