um/mm: remove redundant assignment of max_low_pfn
Current calculation of max_low_pfn is introduced in commit af84eab20891
("[PATCH] uml: fix LVM crash").  It is intended to set max_low_pfn to the
same value as max_pfn.

But I am not sure why the max_pfn is set to totalram_pages, which
represents the number of usable pages in system instead of an absolute
page frame number.  (The change history stops there.)

While we have already calculate it in setup_physmem(), so not necessary to
do it again.

Also this would help changing totalram_pages accounting, since we plan to
move the accounting into __free_pages_core().  With this change,
totalram_pages may not represent the total usable pages at this point,
since some pages would be deferred initialized.

Link: https://lkml.kernel.org/r/20240615034150.2958-1-richard.weiyang@gmail.com
Signed-off-by: Wei Yang <richard.weiyang@gmail.com>
Acked-by: Mike Rapoport (IBM) <rppt@kernel.org>
Cc: Jeff Dike <jdike@linux.intel.com>
Cc: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>
Cc: Alasdair G Kergon <agk@redhat.com>
Cc: David Hildenbrand <david@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>