signal: on exit skip waiting for an ack from the tracer if it is frozen
When a process is being killed or exiting and it has a tracer, it will
notify the tracer and wait for an ack from the tracer to proceed.  However
if the tracer is frozen, this ack will not arrive until the tracer gets
thawed.  This poses a problem especially during memory pressure because
resources of the process are not released.

Things become even more interesting if OOM killer picks such tracee and
adds it into oom_victims.  oom_victims counter will get incremented and
stay that way until tracee exits.  In the meantime, if the system tries to
go into suspend, it will call oom_killer_disable() after freezing all
processes.  That call will fail due to positive oom_victims, but not until
freeze_timeout_msecs passes.  For the whole duration of the
freeze_timeout_msecs (20sec by default) the system will appear
unresponsive.

To fix this problem, skip the ack waiting step in the tracee when it's
exiting and the tracer is frozen.  Per ptrace(2) manual, the tracer cannot
assume that the ptrace-stopped tracee exists.  Therefore this change does
not break any valid assumptions.

Link: https://lkml.kernel.org/r/20240628173247.3507846-1-surenb@google.com
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Debugged-by: Martin Liu <liumartin@google.com>
Debugged-by: Minchan Kim <minchan@google.com>
Cc: Ard Biesheuvel <ardb@kernel.org>
Cc: Christian Brauner <brauner@kernel.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Cc: Tycho Andersen <tandersen@netflix.com>
Cc: Vincent Whitchurch <vincent.whitchurch@axis.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>