mul_u64_u64_div_u64: avoid undefined shift value
Shifting a value to its type's size or beyond is undefined.  This may
happen if the product of a * b is not more than 64 bits despite ilog2(a) +
ilog2(b) being 63 and c having no trailing 0 bits.  We end up with shift=0
and n_lo >> shift | (n_hi << (64 - shift).  Take care of that case and add
such case to the test module.

Using __builtin_ctzll() with 0 is also undefined so take care of that
case too.

Link: https://lkml.kernel.org/r/7rrs9pn1-n266-3013-9q6n-1osp8r8s0rrn@syhkavp.arg
Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
Reported-by: kernel test robot <oliver.sang@intel.com>
Closes: https://lore.kernel.org/oe-lkp/202407121652.69e657c5-oliver.sang@intel.com
Cc: Biju Das <biju.das.jz@bp.renesas.com>
Cc: Uwe Kleine-König <u.kleine-koenig@baylibre.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>